requirements.txt
selenium
Pillow
pywin32
